Provide Docker production build and Kubernetes api prefix
Dockerfile and compose file provided to create a production build of the React components.
The output image is an Nginx web server containing the static assets. This also includes Nginx
configuration for CORS etc. for the Kubernetes API calls.

The /k8s-api prefix is now also required for the kubectl proxy to prevent clashes with the default /
context which is now used by the web application.